backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
sqlparse==0.5.1
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
gunicorn